Renamed `registry` to `functions`.py
